FindingDetail: Extract scoring systems to a separate object
The enum classes representing concrete scoring systems are rather
focused on vulnerabilities. As FindingDetail has now a broader scope,
this is no longer the place to host them. Therefore, move the enums
and the getSeverityString() helper function to a new object
ScoringSystems that combines this functionality in the future.
